removed meter code, moved otel code to profiling
