+++ packages/bsb-core/bsb/__init__.py
@@ -423,8 +423,6 @@ is_partition: "bsb.topology.is_partition"
 is_region: "bsb.topology.is_region"
 load_root_command: "bsb.cli.commands.load_root_command"
 make_config_diagram: "bsb.config.make_config_diagram"
-meter: "bsb.profiling.meter"
-node_meter: "bsb.profiling.node_meter"
 on_main: "bsb.storage.decorators.on_main"
 on_main_until: "bsb.storage.decorators.on_main_until"
 open_storage: "bsb.storage.open_storage"
