+++ packages/bsb-core/bsb/reporting.py
@@ -1,7 +1,3 @@
-import functools
-import importlib.metadata
-import inspect
-import json
 import sys
 import warnings
 
@@ -78,79 +74,6 @@ def warn(message, category=None, stacklevel=2, log_exc=None):
     # Avoid infinite loop looking up verbosity when verbosity option is broken.
     if "Error retrieving option 'verbosity'" in message or options.verbosity > 0:
         warnings.warn(message, category, stacklevel=stacklevel)
-
-
-_otel_tracer = trace.get_tracer("bsb", str(importlib.metadata.version("bsb-core")))
-
-
-def _instrument_command(cls):
-    _orig_handler = cls.handler
-
-    @functools.wraps(cls.handler)
-    def handler(self, context):
-        attributes = {
-            "bsb.type": "command_handler",
-            "bsb.command_class": cls.__name__,
-            "bsb.command_name": cls.name,
-        }
-
-        if hasattr(cls, "get_options"):
-            attributes["bsb.command_options"] = [
-                f"{k}={getattr(context, k)}" for k, v in self.get_options().items()
-            ]
-
-        with _otel_tracer.start_as_current_span(
-            cls.name,
-            attributes=attributes,
-        ):
-            return _orig_handler(self, context)
-
-    cls.handler = handler
-    return cls
-
-
-def _get_implemented_abstracts(cls):
-    """Return dict mapping base class → set of abstract methods implemented in cls."""
-    result = {}
-    subclass_abstracts = getattr(cls, "__abstractmethods__", frozenset())
-    for base in cls.__mro__:
-        if not hasattr(base, "__abstractmethods__"):
-            continue
-        base_abstracts = base.__abstractmethods__
-        # Abstract methods from this base that are now implemented in cls
-        implemented = base_abstracts - subclass_abstracts
-        if implemented:
-            result[base] = implemented
-    return result
-
-
-def _instrument_node(cls):
-    for base, implemented in _get_implemented_abstracts(cls).items():
-        for attr in implemented:
-            orig_method = getattr(cls, attr)
-            if inspect.isfunction(orig_method):
-                wrapped = _make_otel_handler(cls, base, attr, orig_method)
-                setattr(cls, attr, wrapped)
-
-
-def _make_otel_handler(cls, base, attr, orig_method):
-    """Return a wrapper function for a single method."""
-
-    @functools.wraps(orig_method)
-    def handler(self, *args, **kwargs):
-        with _otel_tracer.start_as_current_span(
-            f"{cls.__name__}.{attr}",
-            attributes={
-                "bsb.type": "component_method",
-                "bsb.component_type": base.__name__,
-                "bsb.component_class": cls.__name__,
-                "bsb.component_method": attr,
-                "bsb.component_attributes": json.dumps(self.__tree__()),
-            },
-        ):
-            return orig_method(self, *args, **kwargs)
-
-    return handler
 
 
 __all__ = [
